Checkpointing incomplete work.
There's something wrong with MIDI output, and I'm not sure
when the problem was introduced, so I'm checkpointing this
work.

Includes beginning of implementation of the "info" command,
debugging code for find (which is not working due to aforementioned
problem), and an alternate implementation of MidiSink.

MIDI output is being held *in the MIDI driver* until, I think, the async
runtime shuts down. All the Rust code appears to work correctly,
but actual MIDI output is delayed.