Fixed MIDI syxex output problem.
Turns out that midi-control does *not* provide the EOX terminator
for sysex data. This caused the sysex send to pend inside the Windows
MIDI API. When the output connection is reset, Windows provides the
missing terminator and flushes it out.

Also added a command-line option to adjust the timeout for responses.
This makes debugging much easier.