Port from midi-control to midi-msg.
This attempt was motivated by midi-msg's semantic approach to 14-bit
MIDI values. This is ultimately a dead end because midi-msg generates
"running status" MIDI bytes. OS drivers generally do not accept this
abbreviated form of MIDI for output.

I branched this to preserve the work, but will develop an alternative
approach to dealing with 14-bit values in the main branch.