Added ability to rename WInRT MIDI ports.
First exercise in Rust conditional compilation. The new feature
should exist only in Windows builds.

Switched from midir's default use of WinMM to WinRT because
the Focusrite USB driver misbehaves with WinMM. Getting
extraneous MIDI output after sysexes.

This also contains code for choosing a preset on the device,
but that does not appear to be working yet.